fix: temporarily drop jellyfin-media-player
JMP depends on qtwebengine 5.X which is marked as insecure.
There seems to be some work in the JMP repo to port things to Qt6, but
so far there doesn't appear to be any official releases with that,
despite the fact that Qt5 is EOL and being dropped from various distro's
repos.

I will drop the package for now. If they don't provide a release soon
I'll have to see whether I want to allow the insecure package or use
something else to watch stuff.

Related issues and PRs:
- https://github.com/nixos/nixpkgs/issues/437865
- https://github.com/jellyfin/jellyfin-media-player/issues/282
- https://github.com/jellyfin/jellyfin-media-player/pull/599
- https://github.com/jellyfin/jellyfin-media-player/pull/844

Signed-off-by: Linda Siemons <linda@catbrained.dev>